pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sqlxport"
version = "0.1.3"
description = "Export SQL query results to Parquet and CSV... and upload to S3 or MinIO"
authors = [
    { name = "Vahid Saber" }
]
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }

dependencies = [
    "click",
    "psycopg[binary]",
    "pandas",
    "pyarrow",
    "boto3",
    "python-dotenv",
    "tabulate"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "click",
    "python-dotenv",
    "boto3",
    "duckdb"
]

[project.scripts]
sqlxport = "sqlxport.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]

